fix(cat-gateway): refactor code into health utilities
