+++ catalyst-gateway/bin/src/service/mod.rs
@@ -10,7 +10,7 @@ mod poem_service;
 pub(crate) mod utilities;
 
 use api::mk_api;
-pub(crate) use api::{get_live_counter, set_live_counter, set_not_live, started};
+pub(crate) use api::started;
 pub(crate) use poem_service::get_app_docs;
 use serde_json::{json, Value};
 use tracing::error;
