+++ catalyst-gateway/bin/src/service/utilities/mod.rs
@@ -1,6 +1,7 @@
 //! `API` Utility operations
 pub(crate) mod catch_panic;
 pub(crate) mod convert;
+pub(crate) mod health;
 pub(crate) mod json;
 pub(crate) mod middleware;
 pub(crate) mod net;
